+++ app/src/main/java/org/autojs/autojs/core/ui/JsViewHelper.kt
@@ -6,7 +6,7 @@ import org.autojs.autojs.core.ui.inflater.util.Ids
 
 /**
  * Created by Stardust on 2017/5/14.
- * Transformed by 抠脚本人 on Jul 10, 2023.
+ * Transformed by 抠脚本人 (https://github.com/little-alei) on Jul 10, 2023.
  */
 object JsViewHelper {
     @JvmStatic