+++ app/src/main/java/org/autojs/autojs/core/ui/widget/JsTextViewLegacy.kt
@@ -8,7 +8,7 @@ import androidx.appcompatlegacy.widget.AppCompatTextView
  * Created by SuperMonster003 on Mar 20, 2022.
  * Transformed by SuperMonster003 on May 22, 2023.
  */
-// @Reference to TonyJiangWJ/Auto.js on Mar 20, 2022
+// @Reference to TonyJiangWJ/Auto.js (https://github.com/TonyJiangWJ/Auto.js) on Mar 20, 2022
 class JsTextViewLegacy : AppCompatTextView {
 
     constructor(context: Context) : super(context)