+++ app/src/main/java/org/autojs/autojs/core/ui/xml/XmlConverter.kt
@@ -83,7 +83,7 @@ object XmlConverter {
                 .map(arrayOf("webview", "web"), JsWebView::class.java.name)
                 .map(
                     "text",
-                    // @Reference to TonyJiangWJ/Auto.js on Mar 20, 2022
+                    // @Reference to TonyJiangWJ/Auto.js (https://github.com/TonyJiangWJ/Auto.js) on Mar 20, 2022
                     when (Build.VERSION.SDK_INT < Build.VERSION_CODES.O) {
                         true -> JsTextViewLegacy::class.java.name
                         else -> JsTextView::class.java.name