+++ app/src/main/java/org/autojs/autojs/runtime/api/Timers.java
@@ -1,6 +1,9 @@
 package org.autojs.autojs.runtime.api;
 
 import android.os.Looper;
+import android.os.SystemClock;
+
+import org.autojs.autojs.concurrent.VolatileBox;
 import org.autojs.autojs.core.looper.Timer;
 import org.autojs.autojs.core.looper.TimerThread;
 import org.autojs.autojs.runtime.ScriptRuntime;
@@ -8,24 +11,26 @@ import org.autojs.autojs.runtime.ScriptRuntime;
 /**
  * Created by Stardust on 2017/7/21.
  */
-
 public class Timers {
 
-    private static final String LOG_TAG = "Timers";
+    private final VolatileBox<Long> mMaxCallbackUptimeMillisForAllThreads = new VolatileBox<>(0L);
+    private final Threads mThreads;
+    private final Timer mMainTimer;
+    private final Timer mUiTimer;
 
-    //private VolatileBox<Long> mMaxCallbackUptimeMillisForAllThreads = new VolatileBox<>(0L);
-    private Threads mThreads;
-    private Timer mUiTimer;
-    private ScriptRuntime mRuntime;
 
     public Timers(ScriptRuntime runtime) {
-        mUiTimer = new Timer(runtime, Looper.getMainLooper());
+        mMainTimer = new Timer(runtime, mMaxCallbackUptimeMillisForAllThreads);
+        mUiTimer = new Timer(runtime, mMaxCallbackUptimeMillisForAllThreads, Looper.getMainLooper());
         mThreads = runtime.threads;
-        mRuntime = runtime;
     }
 
     public Timer getMainTimer() {
-        return mRuntime.loopers.getMTimer();
+        return mMainTimer;
+    }
+
+    VolatileBox<Long> getMaxCallbackUptimeMillisForAllThreads() {
+        return mMaxCallbackUptimeMillisForAllThreads;
     }
 
     public Timer getTimerForCurrentThread() {
@@ -34,25 +39,17 @@ public class Timers {
 
     public Timer getTimerForThread(Thread thread) {
         if (thread == mThreads.getMainThread()) {
-            return mRuntime.loopers.getMTimer();
+            return mMainTimer;
         }
         Timer timer = TimerThread.getTimerForThread(thread);
         if (timer == null && Looper.myLooper() == Looper.getMainLooper()) {
             return mUiTimer;
         }
-        if (timer == null) {
-            return mRuntime.loopers.getMTimer();
-        } else {
-            return timer;
-        }
+        return timer;
     }
 
     public int setTimeout(Object callback, long delay, Object... args) {
         return getTimerForCurrentThread().setTimeout(callback, delay, args);
-    }
-
-    public int setTimeout(Object callback) {
-        return setTimeout(callback, 1);
     }
 
     public boolean clearTimeout(int id) {
@@ -61,10 +58,6 @@ public class Timers {
 
     public int setInterval(Object listener, long interval, Object... args) {
         return getTimerForCurrentThread().setInterval(listener, interval, args);
-    }
-
-    public int setInterval(Object listener) {
-        return setInterval(listener, 1);
     }
 
     public boolean clearInterval(int id) {
@@ -79,8 +72,17 @@ public class Timers {
         return getTimerForCurrentThread().clearImmediate(id);
     }
 
+    public boolean hasPendingCallbacks() {
+        // 如果是脚本主线程，则检查所有子线程中的定时回调。mFutureCallbackUptimeMillis用来记录所有子线程中定时最久的一个。
+        if (mThreads.getMainThread() == Thread.currentThread()) {
+            return mMaxCallbackUptimeMillisForAllThreads.get() > SystemClock.uptimeMillis();
+        }
+        // 否则检查当前线程的定时回调
+        return getTimerForCurrentThread().hasPendingCallbacks();
+    }
+
     public void recycle() {
-        mRuntime.loopers.getMTimer().removeAllCallbacks();
+        mMainTimer.removeAllCallbacks();
     }
 
 }